fixes to make linux version compile with new packet counters

+++ src/spindump_util.h
@@ -63,7 +63,7 @@
 //
 
 typedef uint32_t spindump_counter_32bit;
-typedef uint64_t spindump_counter_64bit;
+typedef unsigned long long spindump_counter_64bit;
 typedef struct sockaddr_storage spindump_address;
 typedef struct {
   spindump_address address;
